refactor: extract AgentEvaluationResultStoreMixin from PlaybookMixin (sqlite + ABC)
Move the 7 agent-success-evaluation-result CRUD methods out of the shrinking
PlaybookMixin into a dedicated AgentEvaluationResultStoreMixin, verbatim, for
the sqlite backend and the storage_base ABC. Compose the new mixin into
SQLiteStorage and BaseStorage immediately before PlaybookMixin. Final playbook
sub-mixin extraction: all 56 playbook methods now live in sub-mixins.

Behavior-preserving move; shared module-level helpers stay put and are imported.
