+++ reflexio/server/services/storage/storage_base/playbook/__init__.py
@@ -1,9 +1,11 @@
 from ._agent import AgentPlaybookStoreMixin
+from ._eval_results import AgentEvaluationResultStoreMixin
 from ._optimization import OptimizationJobStoreMixin
 from ._source_linkage import PlaybookSourceLinkageMixin
 from ._user import UserPlaybookStoreMixin
 
 __all__ = [
+    "AgentEvaluationResultStoreMixin",
     "AgentPlaybookStoreMixin",
     "OptimizationJobStoreMixin",
     "PlaybookSourceLinkageMixin",
